Refactor import paths and update error handling in HMM training modules
