+++ src/training/steps/model_training/step09_hmm_based_training.py
@@ -8,16 +8,18 @@ method and profit-based feature engineering, with regime-specific optimization.
 import os
 import warnings
 
-from .core.decorators import handles_errors
-
+from ....core.decorators import handles_errors
+from ....utils.common_operations import (
     get_current_datetime, format_datetime, ensure_directory,
     safe_read_parquet, safe_to_parquet, safe_copy
 )
+from sklearn.feature_selection import (
     f_classif,
     f_regression,
     mutual_info_classif,
     mutual_info_regression,
 )
+from sklearn.metrics import (
     accuracy_score, f1_score, precision_score, recall_score,
     mean_squared_error, mean_absolute_error, r2_score,
     average_precision_score
@@ -27,10 +29,7 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.calibration import CalibratedClassifierCV
 
 # Multi-output training will be imported when needed
-from src.training.steps.step04_analyst_labeling_feature_engineering_components.profit_based_feature_engineering import (
-    ProfitBasedFeatureEngineering
-)
-from .tactician.sr_breakout_predictor import SRBreakoutPredictor
+from ....utils.performance_monitor import (
     PerformanceLevel,
     ValidationLevel,
     adaptive_resource_allocation,
@@ -42,8 +41,8 @@ from .tactician.sr_breakout_predictor import SRBreakoutPredictor
     pipeline_checkpoint,
     validate_feature_engineering_with_lookahead_bias_detection
 )
-from .utils.logger import system_logger
-from .utils.common_operations import ensure_directory, safe_json_dump
+from ....utils.logger import system_logger
+from ....utils.common_operations import ensure_directory, safe_json_dump
 
 # Suppress warnings
 warnings.filterwarnings("ignore")
